refactor: :recycle: refactor popover files
