+++ packages/workspace-fusion/src/modules/garden/components/wrapper/GardenWrapper.tsx
@@ -12,6 +12,7 @@ import { tokens } from '@equinor/eds-tokens';
 import styled from 'styled-components';
 import { GroupingSelector } from '../GroupingSelector';
 import { BehaviorSubject, distinctUntilChanged } from 'rxjs';
+import { GardenPopoverItem } from '../workspace-header/GardenViewSettings';
 Icon.add({ more_vertical });
 
 type GardenWrapperProps<
@@ -60,7 +61,7 @@ export const GardenWrapper = <
     const icon: HeaderIcon = {
       Icon: ({ anchor }) => (
         <GardenPopoverItem
-          config={config}
+          config={config as GardenConfig<any, FilterState>}
           filterState={filterState}
           anchor={anchor}
           groupingKeys$={groupingKeys$.current}
@@ -102,84 +103,3 @@ export const GardenWrapper = <
     </div>
   );
 };
-
-const StyledPopoverHeaderLine = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-type GardenPopoverItemProps = {
-  anchor: HTMLElement;
-  groupingKeys$: BehaviorSubject<string[]>;
-  filterState: FilterState;
-  config: GardenConfig<any, any>;
-  setGroupingKeys: (keys: string[]) => void;
-};
-const GardenPopoverItem = ({ anchor, groupingKeys$, setGroupingKeys, config, filterState }: GardenPopoverItemProps) => {
-  const [isOpen, setIsOpen] = useState(false);
-  const pRef = useRef(null);
-  const [groupingKeys, setInternalKeys] = useState<string[]>(groupingKeys$.value);
-
-  useEffect(() => {
-    const sub = groupingKeys$.pipe(distinctUntilChanged()).subscribe((r) => {
-      setInternalKeys(r);
-    });
-    return () => sub.unsubscribe();
-  }, [groupingKeys$]);
-
-  return (
-    <>
-      <Icon
-        name="more_vertical"
-        color={tokens.colors.interactive.primary__resting.hex}
-        ref={pRef}
-        onClick={() => setIsOpen((s) => !s)}
-      />
-      {createPortal(
-        <Popover style={{ height: '450px', width: '300px' }} open={isOpen} anchorEl={pRef.current}>
-          <Popover.Header>
-            <StyledPopoverHeaderLine>
-              <Popover.Title>Garden settings</Popover.Title>
-              <Icon
-                name="close"
-                color={tokens.colors.interactive.primary__resting.hex}
-                onClick={() => setIsOpen(false)}
-              />
-            </StyledPopoverHeaderLine>
-          </Popover.Header>
-          <Popover.Content>
-            <Suspense fallback={<GroupingSelectorLoading />}>
-              <GroupingSelector
-                groupingKeys={groupingKeys}
-                setGroupingKeys={setGroupingKeys}
-                context={filterState}
-                dataSource={config}
-              />
-            </Suspense>
-          </Popover.Content>
-        </Popover>,
-        anchor
-      )}
-    </>
-  );
-};
-
-const GroupingSelectorLoading = () => {
-  return (
-    <StyledLoadingWrapper>
-      <CircularProgress size={48} />
-    </StyledLoadingWrapper>
-  );
-};
-
-export const StyledLoadingWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  gap: 1em;
-  width: 268px;
-  height: 300px;
-  justify-content: center;
-`;
